xattr-tags-property-page: rename caja-xattr-tags-property-page-extension -> caja-xattr-tags-property-page
